env_*: log command output (#69)
There are two instances where we forget to check the output of a command
when it fails. This commit ensures that we log STDOUT/STDERR in the same
way that we do whenever we capture `CombinedOutput`. Specifically, this
commit adds logging for the output when we fail to recursively chmod
the shared directory as part of the cleanup process.
One note here: recursively chmod-ing the shared directory could easily
fail if a container saves a file as root and the CI user is not root. We
might want to forget about chmod-ing and instead do a `rm -rf`, which
will delete files not owned by the user executing the command.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>