env_kind.go: load image into kind cluster (#73)
* env_kind.go: load image into kind cluster

This commit load the image into the kind cluster with
```
kind load docker-image <image-name>
```
We need to do this if the image is not in a public registry.
Honestly, the purpose of the `preLoadImage()` function is not clear to me.

Signed-off-by: leonnicolas <leonloechner@gmx.de>

* pull and load image

Pull only the image if it isn't present on the machine and always load
it into the cluster.

* Apply suggestion from code review

Signed-off-by: leonnicolas <leonloechner@gmx.de>

---------

Signed-off-by: leonnicolas <leonloechner@gmx.de>